+++ src/cpu/interrupts.c
@@ -1,3 +1,5 @@
+#include <stdint.h>
+
 #include "cpu_private.h"
 #include "gbemu/bus.h"
 #include "gbemu/cpu.h"
@@ -10,7 +12,7 @@ static const uint16_t INTR_VECTORS[5] = {
 int do_intr(struct Cpu* cpu, struct Bus* bus, uint8_t intr) {
   cpu->halted = false;
   if (cpu->ime) {
-    cpu->if_reg &= (uint8_t)~(1u << intr);
+    cpu->if_reg &= (uint8_t)~(1U << intr);
     push(cpu, bus, cpu->PC);
     cpu->PC = INTR_VECTORS[intr];
     cpu->ime = false;
@@ -21,9 +23,10 @@ int do_intr(struct Cpu* cpu, struct Bus* bus, uint8_t intr) {
 
 uint8_t cpu_check_interrupts(struct Cpu* cpu, struct Bus* bus) {
   (void)bus;
-  uint8_t pending = cpu->if_reg & cpu->ie_reg & IF_VALID_MASK;
+  uint8_t pending =
+      (uint8_t)((unsigned)cpu->if_reg & (unsigned)cpu->ie_reg & IF_VALID_MASK);
   for (uint8_t intr = 0; intr < 5; intr++) {
-    if (pending & (uint8_t)(1u << intr)) {
+    if (pending & (uint8_t)(1U << intr)) {
       return (uint8_t)do_intr(cpu, bus, intr);
     }
   }
